move WPResponsiveImage to its own dir
