+++ src/components/Image/Image.test.js
@@ -33,3 +33,4 @@ describe( 'Image', () => {
 		expect( wrapper.find( 'ImageAmp' ).props().src ).toEqual( defaultProps.src );
 	} );
 } );
+
