+++ src/components/WPResponsiveImage/WPResponsiveImage.stories.mdx
@@ -1,0 +1,34 @@
+import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
+import WPResponsiveImage from './WPResponsiveImage';
+
+<Meta title="WPResponsiveImage" component={WPResponsiveImage} />
+
+# WPResponsiveImage
+
+A wrapper for the Image component that automatically generates a `srcset` value based on the provided WordPress image URL (`src`) and a range of image widths (`widthRange`).
+
+## Usage guidelines
+
+- **Do** use the `WPResponsiveImage` for an image whose aspect ratio will not change.
+- **Do** supply an image URL that supports resizing via the `w` and `h` URL query parameters.
+- **Do** supply alt text for non-decorative images.
+- **Do not** use the `WPResponsiveImage` for an image whose aspect ratio changes between breakpoints (use an art direction, i.e., `<picture>` solution instead).
+
+## Props
+
+<Canvas>
+	<Story
+		args={{
+			alt: 'The Apple Inc. logo is seen hanging at the entrance to the Apple store on 5th Avenue in New York',
+			aspectRatio: 1,
+			fallbackWidth: 800,
+			src: 'https://cms.qz.com/wp-content/uploads/2020/10/2020-10-12T095922Z_142848263_RC2XGJ9T1J4Y_RTRMADP_3_APPLE-IPHONE-5G.jpg?quality=80&strip=all',
+			widthRange: [ 200, 1600 ],
+		}}
+		name="WPResponsiveImage"
+	>
+		{args => <WPResponsiveImage {...args} />}
+	</Story>
+</Canvas>
+
+<ArgsTable story="WPResponsiveImage" />
